Fattt snake

+++ Pinnacle/src/main/java/com/cfive/pinnacle/entity/Affair.java
@@ -113,7 +113,7 @@ public class Affair implements Serializable {
 
     @TableField("deleted")
     @TableLogic
-    private Integer deleted;
+    private Long deleted;
 
     @TableField("version")
     @Version
